flask_jwt_extended has been added

+++ code/resources/item.py
@@ -1,5 +1,7 @@
+from flask_jwt_extended.utils import get_jwt_identity
+from flask_jwt_extended.view_decorators import fresh_jwt_required, jwt_optional
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_claims
 from models.item import ItemModel
 
 
@@ -14,7 +16,7 @@ class Item(Resource):
                         required=True,
                         help='This field cannot be blank')
 
-    @jwt_required()
+    @jwt_required
     def get(self, name):
         item = ItemModel.find_by_name(name)
 
@@ -23,6 +25,7 @@ class Item(Resource):
 
         return {'message': 'Item not found'}, 404
 
+    @fresh_jwt_required
     def post(self, name):
         if ItemModel.find_by_name(name):
             return {'message': "An item with name '{}' already exists.".format(name)}, 400
@@ -37,7 +40,11 @@ class Item(Resource):
 
         return item.json(), 201
 
+    @jwt_required
     def delete(self, name):
+        claims = get_jwt_claims()
+        if not claims['is_admin']:
+            return {'message': 'Admin privilege required'}, 400
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
@@ -58,5 +65,14 @@ class Item(Resource):
 
 
 class ItemList(Resource):
+    @jwt_optional
     def get(self):
-        return {'items': [item.json() for item in ItemModel.query.all()]}
+        user_id = get_jwt_identity()
+        items = [x.json() for x in ItemModel.find_all()]
+        if user_id:
+            return {'items': items}, 200
+
+        return {
+            'items': [item['name'] for item in items],
+            'message': 'More data available if you log in.'
+        }, 200
